fix: component-izes AllProducts view
