+++ app/store.jsx
@@ -10,8 +10,7 @@ const store = createStore(
   rootReducer,
   composeWithDevTools(
     applyMiddleware(
-      createLogger({collapsed: true}),
-      thunkMiddleware
+      thunkMiddleware, createLogger({collapsed: true})
     )
   )
 )
